requirement.txt
streamlit==1.40.0
openai==1.40.0
httpx<0.28.0

camelot-py==0.11.0     # PDF tables (needs Ghostscript + Tk if not headless) [web:19][web:26]
pdfplumber==0.11.0          # Backup PDF table strategy [web:23]
tabula-py==2.10.0           # Alternative PDF tables (requires Java) [web:23]
pandas==2.2.3

python-docx==1.1.2          # DOCX tables [web:24]

reportlab==4.2.0            # PDF report with tables [web:25][web:28]
python-dotenv==1.0.1
PyPDF2==3.0.1
pytesseract==0.3.13       # OCR wrapper [web:50][web:57]
pdf2image==1.17.0
Flask==3.0.0
Flask-CORS==4.0.0